refactor: move use-catigory-store to fold categories
